// source/frontend/src/models/api/generated/ApiGen_Concepts_PropertyOperation.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Base_CodeType } from './ApiGen_Base_CodeType';
import { ApiGen_Concepts_Property } from './ApiGen_Concepts_Property';

// LINK: @backend/apimodels/Models/Concepts/PropertyOperation/PropertyOperationModel.cs
export interface ApiGen_Concepts_PropertyOperation extends ApiGen_Base_BaseAudit {
  id: number;
  sourcePropertyId: number;
  sourceProperty: ApiGen_Concepts_Property | null;
  destinationPropertyId: number;
  destinationProperty: ApiGen_Concepts_Property | null;
  propertyOperationNo: number;
  propertyOperationTypeCode: ApiGen_Base_CodeType<string> | null;
  operationDt: string | null;
  isDisabled: boolean | null;
}